💄 갤러리 페이지별 UI 변경사항 반영
