+++ src/Router.jsx
@@ -6,7 +6,6 @@ const Login = React.lazy(() => import('./pages/login'));
 const Signup = React.lazy(() => import('./pages/signup'));
 const Schedule = React.lazy(() => import('./pages/calendar/Schedule'));
 const Gallery = React.lazy(() => import('./pages/gallery'));
-const GalleryMain = React.lazy(() => import('./pages/gallery/GalleryMain'));
 const Albums = React.lazy(() => import('./pages/gallery/Albums'));
 const Album = React.lazy(() => import('./pages/gallery/Album'));
 const PostAlbum = React.lazy(() => import('./pages/gallery/PostAlbum'));
@@ -27,11 +26,9 @@ const Router = () => {
       <Route path="/signup" element={<Signup />} />
       <Route path="/login" element={<Login />} />
       <Route path="/schedules/*" element={<Schedule />} />
-      <Route path="/galleries" element={<Gallery />}>
-        <Route index element={<GalleryMain />} />
-        <Route path="/galleries/:scheduleId" element={<Albums />} />
-        <Route path="/galleries/:scheduleId/:albumId" element={<Album />} />
-      </Route>
+      <Route path="/galleries" element={<Gallery />} />
+      <Route path="/galleries/:scheduleId" element={<Albums />} />
+      <Route path="/galleries/:scheduleId/:albumId" element={<Album />} />
       <Route path="/galleries/posts" element={<PostAlbum />} />
       <Route path="/voice-mails" element={<Voicemail />} />
       <Route path="/community" element={<Community />} />
