+++ src/assets/icons.js
@@ -78,12 +78,6 @@ export const IconSelect = () => (
 );
 
 // Gallery
-export const IconFilter = () => (
-  <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <path d="M6.99969 6H16.9997L11.9897 12.3L6.99969 6ZM4.24969 5.61C6.26969 8.2 9.99969 13 9.99969 13V19C9.99969 19.55 10.4497 20 10.9997 20H12.9997C13.5497 20 13.9997 19.55 13.9997 19V13C13.9997 13 17.7197 8.2 19.7397 5.61C20.2497 4.95 19.7797 4 18.9497 4H5.03969C4.20969 4 3.73969 4.95 4.24969 5.61Z" fill="black"/>
-  </svg>
-);
-
 export const IconComment = () => (
   <svg width="15" height="16" viewBox="0 0 15 16" fill="none" xmlns="http://www.w3.org/2000/svg">
     <path d="M13.5 0H1.5C0.675 0 0 0.68088 0 1.51307V15.1307L3 12.1045H13.5C14.325 12.1045 15 11.4236 15 10.5915V1.51307C15 0.68088 14.325 0 13.5 0ZM13.5 10.5915H2.4L1.5 11.4993V1.51307H13.5V10.5915Z" fill="#A6A6A6"/>
